User: Add Search to PublicInterface
